@@ -10,10 +10,10 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^14.4.3",
-    "@types/jest": "^29.2.2",
+    "@types/jest": "^29.2.3",
     "@types/node": "^18.11.9",
     "@types/react": "^18.0.25",
-    "@types/react-dom": "^18.0.8",
+    "@types/react-dom": "^18.0.9",
     "@types/redux": "^3.6.0",
     "@types/uuid": "^8.3.4",
     "axios": "^1.1.3",
@@ -22,7 +22,7 @@
     "react-redux": "^8.0.5",
     "react-scripts": "5.0.1",
     "redux": "^4.2.0",
-    "typescript": "^4.8.4",
+    "typescript": "^4.9.3",
     "uuid": "^9.0.0"
   },
   "scripts": {
